+++ test/unit/console_tool/__init__.py
@@ -1,0 +1,10 @@
+######################################################################
+#
+# File: test/unit/console_tool/__init__.py
+#
+# Copyright 2023 Backblaze Inc. All Rights Reserved.
+#
+# License https://www.backblaze.com/using_b2_code.html
+#
+######################################################################
+"""Tests for the console_tool commands."""