+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-* Add ability to upload from an unbound source (such as standard input or a named pipe)
+* Add ability to upload from an unbound source such as standard input or a named pipe (fixes #152)
 
+### Deprecated
+* Support of `-` as a valid filename in `upload-file` command. In future `-` will be an alias for standard input.
 
 ### Infrastructure
 * Autocomplete integration tests will now work properly even if tested package has not been installed